project_canvas/python_engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pythonocc-core==7.7.2
